Once upon a time in Denmark, there was a shiny kingdom called ITU. The king of ITU, Anders Høst, was sitting proudly on his throne, gazing over his kingdom, wondering how such wonders could ever exist.
Anders' royal councellor, Nils, was sitting right in
front of him, writing strange stories in his git commit
messages.

Currently, the kingdom of ITU had some really, really
annoying problems with their royal alarms. Nils was sure
that it was simply an issue with the setup, but, as Anders
and Nils would soon discover, a much darker secret was
behind this crazyness...

... to be continued